# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: PrestaRapi - Sistema de Gestión de Préstamos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: getLoans
      tags: [loans]
      summary: List loans
      parameters:
        - name: agencyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Request a new loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/approve:
    post:
      operationId: approveLoan
      tags: [loans]
      summary: Approve a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/reject:
    post:
      operationId: rejectLoan
      tags: [loans]
      summary: Reject a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Loan rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/disburse:
    post:
      operationId: disburseLoan
      tags: [loans]
      summary: Disburse a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisburseRequest"
      responses:
        "200":
          description: Loan disbursed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/reverse:
    post:
      operationId: reverseLoan
      tags: [loans]
      summary: Reverse a disbursement (SuperAdmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disbursement reversed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
components:
  schemas:
...
    Loan:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        agencyId:
          type: integer
        agencyName:
          type: string
        promoterId:
          type: integer
        promoterName:
          type: string
        originalCapital:
          type: number
        currentCapital:
          type: number
        rate:
          type: number
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        status:
          type: string
          enum: [pending_approval, approved, pending_disbursement, active, rejected, settled]
        overdueInterest:
          type: number
        currentInterest:
          type: number
        disbursementDate:
          type: string
          format: date
        nextCutDate:
          type: string
          format: date
        disbursementMethod:
          type: string
        disbursementProof:
          type: string
        inDefault:
          type: boolean
        rejectReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - agencyId
        - originalCapital
        - currentCapital
        - rate
        - periodicity
        - status
        - overdueInterest
        - currentInterest
        - inDefault
        - createdAt
    CreateLoanRequest:
      type: object
      properties:
        clientId:
          type: integer
        agencyId:
          type: integer
        amount:
          type: number
        disbursementMethod:
          type: string
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        notes:
          type: string
      required:
        - clientId
        - agencyId
        - amount
        - disbursementMethod
        - periodicity
    DisburseRequest:
      type: object
      properties:
        rate:
          type: number
        disbursementMethod:
          type: string
        proof:
          type: string
          description: Base64 encoded image or URL
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
      required:
        - rate
        - disbursementMethod
        - periodicity
...