# lib/api-zod/src/generated/types/disburseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrestaRapi - Sistema de Gestión de Préstamos
 * OpenAPI spec version: 1.0.0
 */
import type { DisburseRequestPeriodicity } from "./disburseRequestPeriodicity";

export interface DisburseRequest {
  rate: number;
  disbursementMethod: string;
  /** Base64 encoded image or URL */
  proof?: string;
  periodicity: DisburseRequestPeriodicity;
}